python_solution/requirements.txt
redash-api-client==0.3.0
pandas==1.0.0